// template/react-starter-kit/src/.generated/api/model/resetPasswordRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export interface ResetPasswordRequest {
  /** 챌린지 ID */
  challengeId: string;
  /** 검증 토큰 */
  token: string;
  /** 새 비밀번호 */
  newPassword: string;
  /** 새 비밀번호 확인 */
  confirmPassword: string;
}